Use our own xdr_int64() for all 64 bit types.
Oh dear.  It is really difficult for rpc and tirpc implementors to get
64 bit types right. Sigh.

I am tired of fighting broken ti-rpc libraries under linux and others
that do random corruption on xdr_u_quad_t and friends
so lets just override that shit and replace it with our own, non-broken,
implementation of a 64 bit xdr accessor.