READDIR: Bake the file type into the mode we return
NFS reports type in a separate to mode, where mode only contains
the protection bits. This makes it inconvenient to use when porting programs
since under posix the type is supposed to be part of the mode bits (S_IFMT)

When unmarshalling the directory entries into a nfsdirent structure, bake
the S_IF* file type into where the S_IFMT bits would be.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>