nfs-stream.c: a simple test tool to check performance sequential read mode
./nfs-stream <mfs-url> <bytes-per-second-to-read>

This tool is for testing sequential read mode.
It reads 32kbyte one block at a time using the sync API.
The idea is that when a nfs_[p]read() command is executed, we will once we
identify a sequential scan start issuing asynchronous commands in the
background to populate a readahead cache.

During read, we try to service the request immediately out of cache
and return immediately to the application. During read from cache we will also
process any to previous reads and use it to populate the cache. Also, as the
file offset grows, we will discard the already read blocks from the cache
and issue new asynchronous commands at the tail of the cache.

nfs_read() will when the cache is operating behave like :
  nfs_read() {
     ... process any previous reads and populate the cache ...
     ... send async command to replenish the tail of the cache ...
     return data from cache straigh back to application
  }

The test tool will try to read 32kb chunks at a given rate.
Once the test tool manages to service from cache  the read laency should drop
to <<1ms and the majority of the time spent will be in the sleep() between calls.
Once the latency is steady at <<1ms this is an indication that we now
service all reads out of cache instead of the network.
If the read latency remains <<1ms continously, without any spikes, it means that
we have been successfull in maintaining the asycntrhous tasks to replenish the cache.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>