Revert some commits that cause subtle API issues
We can not have a static rpc->inbuf buffer since that will no longer guarantee
that the received buffer is valid for the duration of callbacks.

One of the problems is that if we issue new (sync) RPCs from within a
callback, that will overwrite and invalidate the receive buffer that
we passed to the callback.

Revert "init: do not leak rpc->inbuf"
This reverts commit f7bc4c8bb1271adbca34dca6488493d32fa155eb.

Revert "socket: we have to use memmove in rpc_read_from_socket"
This reverts commit 24429e95b8079aa91e6b6bf32c4bbe9bcd018e24.

Revert "socket: make rpc->inbuf static and simplify receive logic"
This reverts commit 7000a0aa04e4ee1b312dfcc587e94803f54591ca.