NFSv4: Populate functions should return how many operations they did append.
The populate functions will soon start appending varying number of operations.
For examle the case of open(O_TRUNC) where we would append an extra
SETATTR(size=0) at the end of the compound chain.

Thus use the number of ops that the populate function retured for when
we compute the length of the compound chain instead of assuming it is always
a constant.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>