examples/ld_nfs.c: feature: faking uid and gid
add the possiblity of faking the uid and the gid in the nfs context.
uid and gid value are read from env variables LD_NFS_UID and LD_NFS_GID.
This can be useful on "insecure" enabled NFS share to make the server
trust you as a root.

Few lines added to the README to document the feature.