add conditional support for getpwnam()
Some nfsv4 servers return a user/group name in the attributes instead
of a uid/gid number. At least for the "nobody" user.

For those, try to use getpwnam() where available to map the string back
to a uid/gid or return uid/gid==65534 (common value for nobody)
if not available or failing.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>