multithreading/ptread:  gettid() is not available in all versions
gettid() is not guaranteed to be universally available
so replace this with calling the syscall directly for
building on Linux with Multithreading enabled.

Windows already use a native call for this so it should be unaffected
but other unix-like systems with pthread support will need to have
a replacement function here as well.
This is todo for later for non-linux pthread platforms.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>