libnfs-sync: fail all pdus before returning error from wait_for_nfs_reply
In the not-multithreaded case we need to ensure than the pdu's in flight
are cancelled and removed from the out/wait queues before we
return from wait_for_nfs_reply().
This is because the cb_data structures are allocated on the stack and would
fall out of scope once we return.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>