Consistently use higher maxcount for READDIRPLUS
The explanation is the same as in 4a58e6145568fd21ef956ae6ac3d3c37ef209a57,
i.e. the server would gather 8192 bytes (dircount) of directory entries,
fetch attributes for them and try to fit the extended entries into a
8192 bytes (maxcount) response. Obviously the attributes need quite some
additional space, so only a small subset of the information can actually
be sent back to the client, the rest is wasted.

Let's use the typical 8 to 1 ratio to allow more space for the response.

Benchmark with nfs-ls for different directories on a NetApp NFSv3 share:

    Files   Before    After
 -----------------------------------
    6,650   0.28s --> 0.14s  (-50%)
  163,128    4.8s -->  1.7s  (-65%)
  330,602    9.6s -->  3.2s  (-67%)

A significant improvement from ~34k to ~100k entries/s.
Further increasing dircount/maxcount didn't lead to higher performance.