Write multiple pdus at once.
Attempt to batch the iovectors from several pdus in order to reduce the
number of calls to writev.
In the best case this change will reduce the number of calls to writev
to an eighth of what they were before.

The maximum number of vectors was reduced to match the limit set on the
Linux kernel for the fast code path that handles iovectors.