rpc: re-add fragment handling when reading from socket.
It is very inefficient since for fragmented PDUs we allocate a reassembly
buffer we memcpy() all the data into before we process it, and for reads
we later memcpy() once again from this buffer to the application.
More elegant but complex solution would be to keep all fragments in an
array of iovectors and avoid that first memcpy().

But, fragment handling is really only meant for interoperability with
ancient NFS servers so it is not worth the effort to optimize for this case.
Most/all modern NFS clients and server just flat out do not support
fragments at all,  so ...

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>