Reset state on reconnect.
Otherwise we may be attempting to read data belonging to the state of a
previous socket which is now closed.

I could not find any code resetting the state during an automatic
reconnect.

We had a situation where the socket was closed whilst on READ_PAYLOAD
state due to a record marker that specified a payload bigger than
supported and libnfs kept failing and reconnecting without processing
any data because it kept trying to adjust the input buffer to the
unsupported size.