Allow setting a limit on the number of commands in-flight.
Some NFS servers will limit the number of in-flight requests from NFS clients
to a maximum of 128.
This limit is configured as "sunrpc.tcp_max_slot_table_entries".
When a client reaches that limit the server responds by decreasing the
window size for the TCP connection, hence decreasing throughput until
the window size recovers.

The new functionality will allow applications to set a limit on the
number of in-flight requests (or "awaiting" as the API defines them).

Also, a new function was added to request the number of commands waiting
because this number may be very different to the total size of the
output queue and it could be useful to know the difference between the
two.