nfsv3:open make sure path is nul-terminated for relative paths
Allocate strlen(path) + 2 when processing relative paths in
nfs3_initial_open_cb(). We need +2 because we memcpy the path
into &ptr[1] and need an extra nul-byte at the end to terminate
the string.
Otherwise the resulting path might not be terminated properly and
may end up with a garbage trailer.

Reported-by: Amulya Nataraj <Amulya.N@microsoft.com>
Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>