Fix segv for UDP packets (broadcast portmap/callit)
This broke recently in the re-factoring to add zero copy reads
and where we broke out the XID value into a new field in the rpc structure.

Assign the xid value properly when reading a UDP packet and check that
we could match the incoming PDU to a request in flight.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>