If mounting v3 fails, try mounting v4 instead
If we did not specify a specific version of nfs, then first try v3 and
if that fails try v4 instead.
Only do this if no version was specified, IF a version was specified then
only that version will fail and there will be no fall-back to a different
version.

Only implement for the sync API for now.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>