Add an example of multithreading with async calls
Example of using multithreading and the async api to copy a file
to an nfs server.

This example will create 4 service threads that will service
the socket, i.e. reading/writing to the socket (as well as executing the
async callbacks).

The example program then creates multiple worker threads, each handling
a 10MByte section of the file to be copied.
Each thread will immediately issue async pwrite calls for the entire section
and then wait until all the callbacks have completed.
The threads issue the async calls concurrently and will result in an
enormous number of concurrent calls being issued to the server.
Some servers have been known to crash once you go above a few hundred
thousandis of concurrent I/Os.

As these are async calls, the callbacks will execute in the context of the
service context, which means you need to keep the callbacks light-weight
or you may need to increate the number of contexts/service threads.
This is different from when using the sync api where most of the processing
will happen in the worker thread once the responce has been received from
the server.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>