Fix include sys/time.h
POSIX says `struct timeval` is defined if <sys/time.h> is included.

Instead of the mess that is currently done based on the system on which
the stuff is being compiled, always include it except when _WIN32 is
defined.

This will fix the following build failure on musl for any applications
using libnfs (e.g., mpd):

$SYSROOT/usr/include/nfsc/libnfs.h:965:23: error: field 'atime' has incomplete type 'timeval'
        struct timeval atime;

Fix #272

Signed-off-by: Fabrice Fontaine <fontaine.fabrice@gmail.com>