Always set TCP_NODELAY
It can improve performance in some cases.

A common drawback with TCP_NODELAY is that it may cause additional
small segments to be sent and increase the amount of cpu processing.
That is not the case in libnfs as we never write small buffers to the socket
but always coalece everything into a large iovec and then send everything
as a single call to writev()

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>