reorder calling stat_cb and the pdu->cb
so that we always call rpc->stat_cb before calling pdu->cb.
Note that the stat_cb callback and the pdu callback are called from
the same context and the same function, rpc_process_reply(), this guarantees
that when the pdu callback is invoked it is guaranteed that the most recent
rpc->stat_cb invocation is for the same PDU as what teh pdu callback is
invoked for.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>