Change logging to use an application callback instead of writing to stderr
Stderr/filedescriptor 2 does not exist for daemons (fork twice and close
all file descriptors) or embedded systems that might not even have a console
or a filesystem to write to.

This fixes RPC_LOG and LOG. Still todo is decide how to handle
TLS_LOG.
All references to stderr must be removed from the library.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>