Exclude requests timing out in outqueue from causing major recovery (reconnect)
Requests that timeout in outqueue have not been sent to server and hence they
do not signifiy any issue with the server. Yes they may indicate issue with the
server indirectly since requests ahead of them did not get responses, but any
corrective action is taken on behalf of those requests which have been sent to
server.
Also added a new stats member num_timedout_in_outqueue to convey this to application.