Add non-IO RPC PDUs to head of outqueue
This is to avoid huge number of WRITE (and READ) PDUs to hijack a connection
causing commands like ls/stat/find/chmod to appear to hang for a long time,
since WRITEs can take really long to be sent out due to TCP window getting
full (if server is unable to process them fast enough).

POSIX has compliance requirement only for completed requests. Since we are
only reordering ongoing requests (not yet completed) it should not cause any
violation. Moreover I cannot think of any application which will depend on
any ordering guarantees of parallel requests.

Let's monitor it and if it causes any issuesm revisit on a case-by-case basic.