Fix race-condition when sending UDP PDUs when multithreading is enabled
There is a race condition for UDP with multithreading where we could
send a UDP packet from the primary thread but then the response
could be received and processed by the background socket i/o before
we had time to add the PDU/XID to the list of commands in flight,
resulting in a lost response.

Fix this by first adding the PDU to the list before sending.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>