Resolve server address before reconnect. (#28)
Default behavior is same as before, i.e., reconnect to the same address as
was resolved the very first time we connected to the server, but user can
choose the "resolve before reconnect" behaviour by calling
nfs_set_resolve_on_reconnect(). With that set, everytime we have to reconnect
we first resolve the server address and reconnect to that address instead
of the originally resolved address.
This enables support for NFS server migration where the NFS server address
can change after mount. libnfs can simply reconnect to the new address and
start exchanging RPCs with that. There is no need to perform mount with the
new address.

Co-authored-by: Nagendra Tomar <natomar@microsoft.com>