Allow user to find out whether a PDU that completed was retransmitted. (#29)
Was the PDU retransmitted? i.e., did we send it to the server more than
once? User can use this info to relax some errors, f.e., a REMOVE request
that fails with NFS3ERR_NOENT can be treated as success if it was a
retransmited request.
Note that most applications will handle an unlink() call succeeding for a
non-existent file better than unlink() call failing with NOENT for a file
that was actually present.

Co-authored-by: Nagendra Tomar <natomar@microsoft.com>