Do not segfault in process with no controlling terminal
The getlogin() function can return NULL in some cases, one example is when the calling process has no controlling terminal, such as a service process running under systemd. If getlogin() returns NULL this will cause a segfault in libnfs as a NULL string is passed to rpc_set_username, which is then passed to strdup. The getlogin() function is being used to obtain a username for use with krb5 security even if krb5 security is never used. We can gracefully avoid a segfault by falling back to the username for current effective uid in the case that getlogin() returns NULL.

Alternative approaches:
* Allow username to be NULL but check for this when setting up krb5 security and return an error to caller then
* Don't use getlogin() to obtain a username for Kerberos and instead allow the caller to specify a principal, and allow Kerberos to select the default principal when no principal is provided by the user.